Add Radial fill to yard scouting
I added a radial fill image to the Dog Image prefab, and added a few
lines of code to dogcollarslot.cs. It piggybacks on the existings handle
and init methods. I tagged my work with "BP" out of habit. If this is
too much commenting, let me know. The coroutine lerps the image down
each second for a second.